omsa: Remove lazy-lock.json from gitignore
According to the README information in the document, this is something
that should be done once the repository is forked.

This lock file saves the versions of the different installed plugins in
the system. It can be used to restore all the plugins in a given machine
to the version saved in the lock file, ensuring the same behaviour on
different machines.