tmp: omsa: Update noice version
The current version of the noice plugin in the upstream have a bug,
where command-line-code is not being sent to the user. This is quite
big, and have a big effect on the usability of the plugin.

There is a PR open for the fix of that, available here:
https://github.com/folke/noice.nvim/pull/1098

However, until this PR is resolved, the plugin becomes unusable. Instead
of disabling it, we switch the used version to the forked version, which
have the fix available in it.
This commit should be reverted once the PR is solved in the upstreamed
noice.