omsa: gitsigns: Use included file
Update the configuration regarding the usage of gitsigns. Originally,
kickstart provides gitsigns as available by default, but does not
provide keymaps to it. I've added the keymaps manually, before
notificing that kickstart have an available file with these options
already configured.

This commit removes the configuration from the main init.lua file,
instead, using the included file for the provided keymaps.
In addition, it changes the included file to include all (and only) the
keymaps that are relevant to my usage of the plugin.